First attempt at the basis transformation for conforming Arnold-Winther.
