+++ finat/__init__.py
@@ -10,7 +10,8 @@ from .bell import Bell                  # noqa: F401
 from .hermite import Hermite            # noqa: F401
 from .mtw import MardalTaiWinther       # noqa: F401
 from .morley import Morley              # noqa: F401
-from .aaw import ArnoldAwanouWinther    # noqa: F401
+from .aw import ArnoldWinther           # noqa: F401
+from .aw import ArnoldWintherNC         # noqa: F401
 from .trace import HDivTrace  # noqa: F401
 from .spectral import GaussLobattoLegendre, GaussLegendre  # noqa: F401
 from .tensorfiniteelement import TensorFiniteElement  # noqa: F401
